Pokemon Initialization 50%

+++ Assets/Scripts/SystemManager.cs
@@ -1,15 +1,19 @@
-﻿using System.Collections;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEngine.UI;
 using System;
+using System.Collections;
 using System.IO;
+using System.IO.IsolatedStorage;
 using System.Runtime.Serialization.Formatters.Binary;
 
 public class SystemManager : MonoBehaviour 
 {
+    //INI reading
+    string[] contents;      //Contents of the ini
+    bool processing = false;//Whether INI is currently being read
+
 	//Error logging
 	string errorLog;		//Location of error log
-	string errorData;		//The details of the error
 	StreamWriter output;	//The error writer
 
 	//Persistent variables
@@ -28,6 +32,72 @@ public class SystemManager : MonoBehaviour
 	GameObject arrow;		//Arrow to signal end of text
 	bool displaying;		//If text is currently being output
 
+    #region INIReading
+    public void GetContents(string iniLocation)
+    {
+        if (!processing)
+        {
+            processing = true;
+            contents = File.ReadAllLines(iniLocation);
+            processing = false;
+        } //end if
+    } //end GetContents(string iniLocation)
+
+    public T ReadINI<T>(string section, string key)
+    {
+        //System.Diagnostics.Stopwatch myStopwatch = new System.Diagnostics.Stopwatch ();
+        //myStopwatch.Start ();
+        object value = null;
+        string sectionName = "[" + section + "]";
+        for (int i = 0; i < contents.Length; i++)
+        {
+            //Check for section
+            if (contents[i] == sectionName)
+            {
+                //Search through section
+                while(true)
+                {
+                    //Increment and break if line starts with bracket
+                    i++;
+                    if(i >= contents.Length || contents[i].StartsWith("["))
+                    {
+                        break;
+                    } //end if
+
+                    //Check for key
+                    if (contents[i].StartsWith((key)))
+                    {
+                        //Break line at delimiter, return 2nd part, break loop
+                        string[] result = contents[i].Split ('=');
+                        value = result [1];
+                        break;
+                    } //end if
+                } //end while
+
+                if(value == null)
+                {
+                    value = default(T);
+                } //end if
+                
+                //Section finished search
+                break;
+            } //end if
+        } //end for
+       // myStopwatch.Stop ();
+       
+       // Debug.Log (myStopwatch.ElapsedMilliseconds + " elapsed");
+
+        //Return the information in the type requested
+        return (T)Convert.ChangeType (value, typeof(T));
+    } //end ReadINI(string section, string key)
+
+    //Returns all parts in a CSV line
+    public string[] ReadCSV(int section)
+    {
+        return contents [section].Split(',');
+    } //end ReadCSV(int section)
+    #endregion
+
 	#region ErrorLog
 	//Initialize the error log
 	public bool InitErrorLog()
@@ -35,28 +105,31 @@ public class SystemManager : MonoBehaviour
 		//Init errorLog location
 		errorLog = Environment.GetEnvironmentVariable ("USERPROFILE") + "/Saved Games/Pokemon Carousel/error.log";
 
-		//Append data if a file already exists
-		try
-		{
-			output = new StreamWriter (errorLog, true);
-		} //end try
-		//Manage exception
-		catch(System.Exception ex)
-		{
-			//If directory doesn't exist
-			if(ex.GetType() == typeof(DirectoryNotFoundException))
-			{
-				//Create directory
-				System.IO.Directory.CreateDirectory(Environment.GetEnvironmentVariable("USERPROFILE") + 
-				                                    "/Saved Games/Pokemon Carousel");
+        if (Directory.Exists (Environment.GetEnvironmentVariable ("USERPROFILE") + "/Saved Games/Pokemon Carousel"))
+        {
+            //Create or get errorLog
+            output = new StreamWriter (errorLog, true);
+        } //end if
+        else
+        {
+            try
+            {
+                //Create directory
+                Directory.CreateDirectory(Environment.GetEnvironmentVariable ("USERPROFILE") + 
+                                          "/Saved Games/Pokemon Carousel");
 
-				//Create error log
-				output = new StreamWriter (errorLog, true);
-			} //end if
-		} //end catch
+                //Create or get errorLog
+                output = new StreamWriter (errorLog, true);
+            } //end try
+            catch(SystemException ex)
+            {
+                Debug.LogError("Could not create directory because " + ex.ToString());
+            } //end catch
+        } //end else
+
 		//Send a starting line
 		LogErrorMessage (DateTime.Now.ToString () + " - Game was started.");
-
+       
 		//Report whether successful
 		if(output != null)
 		{
@@ -80,16 +153,15 @@ public class SystemManager : MonoBehaviour
 		} //end if
 	} //end LogErrorMessage(string message)
 
-	//Closes error log
-	void CloseErrorLog()
-	{
-		//Make sure an error log exists
-		if(output != null)
-		{
-			output.Close();
-			output = null;
-		} //end if
-	} //end CloseErrorLog
+    //Closes error log file
+    void OnApplicationQuit()
+    {
+        if (output != null)
+        {
+            output.Close();
+            output = null;
+        } //end if
+    } //end OnApplicationQuit
 	#endregion
 
 	#region Text
@@ -319,6 +391,5 @@ class PersistentSystem
 	public int pBadges;
 	public int pHours;
 	public int pMinutes;
-	public int pSeconds;
-    public Pokemon aPokemon;
+    public int pSeconds;
 } //end PersistentSystem class
